+++ CHANGES.md
@@ -2,6 +2,8 @@
 * Remove support for Observable methods.
 * SaxDecoder now decodes multiple types.
 * Remove pattern decoders in favor of SaxDecoder.
+* Use single non-generic Decoder/Encoder instead of sets of type-specific Decoders/Encoders.
+* Decoders/Encoders are now more flexible, having access to the Response/RequestTemplate respectively.
 
 ### Version 4.4.1
 * Fix NullPointerException on calling equals and hashCode.